feat(frontend): full-screen tabbed Guild Settings overlay
Move guild settings out of the cramped 320px top-bar popover into a
side-tabbed overlay panel (General / Foreman / Claude / Spawn Defaults
/ Members). Env vars now live on a full-height Foreman tab with an
"inherited by every worker" note instead of being buried behind a
Configure collapse. Esc and backdrop-click close it.

TopBar slims down to just the gear trigger; all settings logic moves
into the new GuildSettingsPanel component. No backend changes.
